refactor the code logic
